Create routes folder
