+++ src/lib/routes/NotFound.js
@@ -1,0 +1,3 @@
+import React from 'react'
+
+export default (slug) => (<p>Project not found: {slug}</p>)
