--- src/lib/util/handlers/errors/index.js
@@ -1,4 +1,0 @@
-import React from 'react'
-
-export const notFound = (slug) => (<p>Project not found: {slug}</p>)
-export const error    = ()     => (<p>Error</p>)
